requirements.txt
numpy==1.26.4
pandas>=2.2,<3
scikit-learn>=1.5,<2
matplotlib>=3.8,<4
scikit-surprise==1.1.4
seaborn>=0.13,<1